pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "karya"
version = "0.1.0"
description = "A task manager that honors your work"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Shridhar S Sherugar", email = "shridhars145@gmail.com"},
]
dependencies = [
    "click>=8.2.1",
    "colorama>=0.4.6",
    "tabulate>=0.9.0",
    "python-dateutil",
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Scheduling",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

keywords = ["task", "todo", "cli", "productivity", "task-manager"]

[project.scripts]
karya = "karya.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/karya"
Issues = "https://github.com/yourusername/karya/issues"
Repository = "https://github.com/yourusername/karya"